feat: Refactoring ProfileHeader for constructive abstraction
