+++ src/components/modules/profile/components/ProfileHeader/components/index.ts
@@ -1,0 +1,3 @@
+export * from './ActionButtons';
+export * from './ContactInformation';
+export * from './PDFGenerator';
