+++ src/components/modules/profile/components/ProfileHeader/index.ts
@@ -1,2 +1,4 @@
 export { default as ProfileHeader } from './ProfileHeader';
 export type { ProfileHeaderProps } from './types';
+export * from './components';
+export * from './utils';
