+++ src/components/modules/profile/components/ProfileHeader/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './dateFormatters';
+export * from './phoneFormatter';
